Fixed a few more warnings about type errors

+++ src/dataframe/schema.h
@@ -61,7 +61,7 @@ public:
 		types_ = new DataType[capCols_];
 		if (from.width() != 0)
 		{
-			for (int i = 0; i < from.width(); i++)
+			for (size_t i = 0; i < from.width(); i++)
 			{
 				add_column(from.col_type(i));
 			}
